fastapi_project/requirements.txt
fastapi
uvicorn
requests
python-multipart
huggingface-hub
python-dotenv
google-api-python-client
google-generativeai
pandas
